refactor: extract transaction and evidence services
